Add files via upload
This is the week1 submission of the project of Water Quality Prediction. Here I have worked to check the pollutants and impurities in the water and safe to drink and use or not. I am using jupyter to code for the project. I am using python as the language to code. 
Here in this project, I have imported MultiOutputRegressor from MachineLearning library. And the algorithm which I have used is ensemble method.
We always need to know the number of rows and columns so we use df.shape to know about it. To know the information about dataset we use df.info . 
I have taken the pollutants Co2, NO3, NO2, SO4, PO4.